Add Makefile to compile the project
You can compile both the client and the daemon with `make all`,
only the client with `make cassini`, or only the daemon with
`make saturnd`.
You can also remove all the executable and object files with
`make clean`.

To add a file to either the client or the daemon, follow the
instructions at the top of the Makefile.